app/src/main/java/com/mawumbo/storyapp/ui/home/HomeViewModel.kt
package com.mawumbo.storyapp.ui.home

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.mawumbo.storyapp.data.repository.StoryRepository
import com.mawumbo.storyapp.model.Story
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class HomeViewModel @Inject constructor(
    private val repository: StoryRepository
) : ViewModel() {

    val allStory: LiveData<PagingData<Story>> = repository.getAllStory()
        .cachedIn(viewModelScope)

    fun logout() {
        viewModelScope.launch {
            repository.logout()
        }
    }

}
